requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-dotenv>=1.0
pydantic>=2.0
